Status Invest Scrapers
These Python Scripts are capable to do a Web Scrap to collect the CNPJ.

Each type of Stock negociated at B3 Stock Market have a specific script.

They are:
 - Ações
 - FIIs = Fundos Imobiliários
 - ETF = Exchange Traded Fund

This was done for now because each of them has different page structures

The XPath of each one are different.

In future versions we may do a Refactoring